Add a simple Title functionality
This can be made more customizable (like allowing the user to set a custom title for each alert type), if of any interest.
The only issue that I have is that I am not a coffee user, and I rather prefer doing it directly in js or jquery. I need to use a translator in order to bring my code in coffee.